Made root a field of EveryScreen, removed from createUI.
This enables smoother handling of async responses. We can issue a request for
data as soon as the screen is created, but then defer the creation of the UI
for the response until the screen show transition is complete (which ensures
that we don't jank up the transition animation by starting to create a ton of
UI elements while it's running).

Did this unjanking for the NewsScreen, other screens to follow as needed.