A shitton of work on fancy collection screen animations.
It's pretty good now, but there are sometimes edge cases which are hard to deal
with. Maybe there's a better approach to smooth expandy contracty lists that
would somehow magically deal with these edge cases, but I'll have to think
about it. The approach I settled on is way simpler/more robust than a few early
ones I tried.