A bunch of powerup menu shenanigans.
We now base pup menu enabledness on whether there are actually unflipped cards
left, not on a half-assed interpretation of free flips and next flip cost
(which would be wrong if you flipped your whole grid and still had free flips).

We also fade out the pup button when disabled but don't hide it, so that the UI
doesn't jerk up due to the removal of the taller pup menu button.

Lastly, the next flip cost and pup button are in separate hboxes each of which
are given half the screen. This avoids relayout herky jerky when the next flip
cost changes and triggers a relayout.