Revamped gesture handling. Improved card swiping.
The new gesture handling allows us to directly manipulate the card. I still
need to redo things to compute the swipe velocity and animate the card based on
the velocity imparted to it by the swipe, but things are now in a better
position to do that sort of thing.