Display front of card based on ThingCardPlus only.
When viewing a full series, we have enough info from the getSeries result to
display a card's front, so we defer the loading of the card details
until/unless they flip the card over. This makes flipping through a series much
snappier.

I need to rethink the giver info. I'm probably just going to move it onto the
card back, though that can be cramped if the card details are long... blah.