Use a bullet instead of an arrow. Way simpler.
The fucking arrow is not available in any of Android's stock fonts, and of
course it's never available in any custom fonts. Anyway, switching fonts for
the auto-shrinking path label was a bunch of PITA anyway. Now we just use the
bullet that's already in the handwriting font and it becomes a one-liner.

Fonts... Sigh.