Pass the game into buyProduct, enumerate skus.
There's a bit of a Gordion knot re: game and device, so passing game into
buyProduct simplifies things.

Enumerating the SKUs is necessary because Google provides no way to get a list
of SKUs you have on sale, so you have to know in advance, yay! It's not like
they've changed over the five year lifetime of the game, so whatevs.