Updated to latest FB Android SDK (3.5).
This allows us to stop using the legacy FB API for showDialog.

This is also untested because of some annoying bullshit. The maven-scala-plugin
requires that all code (.java and .scala) be compiled via scalac, due to
integration with the zinc incremental compiler (which we're not using, but we
can't tell the plugin to split compilation between javac and scalac), but
scalac chokes on the .java code shipped with the FB SDK because it declares a
private base class with public methods that are intended to be called via a
public class that derives from the private base class, but scalac thinks that's
totally uncool and refuses to compile it (even though it's legal, if dubious,
in Java).

It's a miracle that any real work gets done with so many fucking moving parts.

Anyway, now I'll probably have to go rewrite the whole Android backend code in
Java. That's not a huge deal since it's only 384 lines of Scala code, but sigh.