Force reauth with Facebook if server reports FB error.
If we send cached FB credentials to the server and it chokes on them, we need
to force a reauth with Facebook to get fresh credentials before trying again.

This also introduces better failure handling for session validation. Instead of
just choking and dying and leaving the app unusable, we offer to "Try again".