Bump to 3.9.0 Facebook SDK.
This revamps the Open Graph stuff to in theory be the way it's supposed to be
for the Xamarin bindings to Facebook's iOS protocol bullshit, but of course it
still doesn't work. It seems like maybe there's a type missing in the OG
action, but there's no way to supply it via the APIs. Fuck if I know.

I'm probably just going to tear out Facebook sharing entirely (on iOS, it works
on Android) because I've wasted waaaaaaay too much time on it and no one is
ever going to use it, and even if they did use it, it will never show up in
anyone else's feed because Facebook buries this shit anyway.